added controllers directory and restructed routes
