+++ routes/admin/scholar.js
@@ -1,20 +1,18 @@
 const express = require('express');
 const router = express.Router();
 const catchAsync = require('../../utils/catchAsync');
-const Scholar = require('../../models/scholar');
 
-router.get('/', catchAsync(async (req, res) => {
-    const scholars = await Scholar.find({});
-    res.render('templates/admin/scholars/index', { scholars });
-}))
-router.get('/:id', catchAsync(async (req, res) => {
-    const { id } = req.params;
-    const scholar = await Scholar.findById(id);
-    // res.send("Gor request");
-    res.render('templates/admin/scholars/show', { scholar });
-}))
+const scholars = require('../../controllers/admin/scholar');
+
+router.get('/', catchAsync(scholars.indexPage));
+
+router.get('/:id', catchAsync(scholars.showPage));
+
 module.exports = router;
 
+
+//CODE THAT IS NOT NEEDE RIGHT NOW.BUT MIGHT BE USED LATER
 // const ExpressError = require('../../utils/ExpressError');
 // const passport = require('passport');
 // const User = require('../../models/user');
+// const Scholar = require('../../models/scholar');
