+++ views/templates/applicationForm/courseDetails.ejs
@@ -30,7 +30,8 @@
         </div>
         <div class="row">
             <div class="col-md-6 mb-3">
-                <label for="researchTopic" class="form-label">Research Topic</label>
+                <label for="researchTopic" class="form-label">Research Topic <span
+                        class="required-field">*</span></label>
                 <input type="text" class="form-control" id="researchTopic" name="scholar[topic]">
             </div>
         </div>
