+++ views/templates/loginForm.ejs
@@ -1,6 +1,6 @@
 <% layout('/layout/boilerplate') %>
     <div class="row justify-content-center">
-        <div class="col-8 col-md-6">
+        <div class="col-8 col-md-5 col-lg-4">
             <form action="/login" method="POST" class="validated-form" novalidate>
                 <div class="mb-3">
                     <label for="username" class="form-label">Username</label>
