+++ views/templates/registerForm.ejs
@@ -1,6 +1,6 @@
 <% layout('/layout/boilerplate') %>
     <div class="row justify-content-center">
-        <div class="col-8 col-md-6">
+        <div class="col-8 col-md-4">
             <form action="<%= originalPath %>" class="validated-form" method="POST" novalidate>
                 <div class="mb-3">
                     <label for="username" class="form-label">Username</label>
